circleCI memory overload bug
